block: anxiety-iosched: use uint8_t for starvation limit
We don't want user's setting the starvation limit to a ridiculous
number. This should generate bounds of 0-128, which should be plenty.

+++ block/anxiety-iosched.c
@@ -10,7 +10,7 @@
 #include <linux/init.h>
 
 /* default tunable values */
-static const unsigned int max_writes_starved = 12; /* max amount of times reads can starve pending writes */
+static const uint8_t max_writes_starved = 12; /* max amount of times reads can starve pending writes */
 
 struct anxiety_data {
 	struct list_head queue[2];
